+++ src/api/news/types.ts
@@ -22,9 +22,3 @@ export type ArticleItem = {
   content: string;
   pictureUrl: string;
 };
-
-export type HttpErrorType = {
-  name: string;
-  message: string;
-  statusCode: number;
-};
