+++ src/errors/http/types.ts
@@ -1,0 +1,5 @@
+export type HttpErrorType = {
+  name: string;
+  message: string;
+  statusCode: number;
+};
